Enhance setup.sh: handle if .gitconfig 404 + configure git user name/email

+++ config/setup.sh
@@ -39,11 +39,14 @@ echo "INFO: creating '~/.ssh/cm_socket' if it doesn't exist"
 mkdir -p ~/.ssh/cm_socket
 
 echo "INFO: updating the global git config"
-cp -v ~/.gitconfig ~/.gitconfig.backup.$IDEMPOTENTIFY_MARK.$(date +%s)
+# Skip cp if file doesn't exist (set -e is in effect)
+[[ -e ~/.gitconfig ]] && cp -v ~/.gitconfig ~/.gitconfig.backup.$IDEMPOTENTIFY_MARK.$(date +%s)
 git config --global core.editor "vim -u $SRC_DOTFILES_DIR/workman.basic.vimrc -c 'set nomodeline'"
 git config --global receive.denyCurrentBranch updateInstead
 git config --global core.sshCommand "ssh -F $SRC_DOTFILES_DIR/ssh_config"
 git config --global init.defaultBranch master
+git config --global user.name "ps"
+git config --global user.email "24826492+therootusr@users.noreply.github.com"
 
 echo "INFO: saving existing zshrc"
 cp -v $ZSHRC_PATH $ZSHRC_PATH.before.$IDEMPOTENTIFY_MARK.zshrc
